requirements-dev.txt
pytest
pytest-flask
pytest-cov
pytest-mock
black
flake8
mypy
